requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==5.0.4
django-bootstrap5==24.2
django-crispy-forms==2.1
django-easy-audit==1.3.6
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
python-decouple==3.8
python-dotenv==1.0.1
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
